refactor: move create-node CLI command and helper functions to commands/ folder
